Use shared-libraries example
